requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Rate limiting
slowapi==0.1.9

# Métricas e monitoramento
prometheus-client==0.19.0
structlog==23.2.0

# Logging estruturado
structlog==23.2.0

# Validação de dados
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2

# Redis (opcional para rate limiting)
redis==5.0.1

# Banco de dados
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Utilitários
python-dotenv==1.0.0 